refactor: move from_* helpers to types.document.generators
