+++ tests/unit/clients/python/test_client.py
@@ -12,6 +12,7 @@ from jina.flow import Flow
 from jina.parsers import set_gateway_parser, set_client_cli_parser
 from jina.peapods import Pea
 from jina.proto.jina_pb2 import DocumentProto
+from jina.types.document.generators import from_csv, from_ndjson, from_files
 
 cur_dir = os.path.dirname(os.path.abspath(__file__))
 
@@ -94,7 +95,7 @@ def test_mime_type(restful):
             assert d.mime_type == 'text/x-python'
 
     with f:
-        f.index(DocumentArray.from_files('*.py'), validate_mime_type)
+        f.index(from_files('*.py'), validate_mime_type)
 
 
 @pytest.mark.parametrize('func_name', ['index', 'search'])
@@ -104,7 +105,7 @@ def test_client_ndjson(restful, mocker, func_name):
         os.path.join(cur_dir, 'docs.jsonlines')
     ) as fp:
         mock = mocker.Mock()
-        getattr(f, f'{func_name}')(DocumentArray.from_ndjson(fp), on_done=mock)
+        getattr(f, f'{func_name}')(from_ndjson(fp), on_done=mock)
         mock.assert_called_once()
 
 
@@ -115,7 +116,7 @@ def test_client_csv(restful, mocker, func_name):
         os.path.join(cur_dir, 'docs.csv')
     ) as fp:
         mock = mocker.Mock()
-        getattr(f, f'{func_name}')(DocumentArray.from_csv(fp), on_done=mock)
+        getattr(f, f'{func_name}')(from_csv(fp), on_done=mock)
         mock.assert_called_once()
