+++ tests/unit/flow/test_asyncflow.py
@@ -4,9 +4,10 @@ import time
 import numpy as np
 import pytest
 
-from jina import Document, DocumentArray
+from jina import Document
 from jina.flow.asyncio import AsyncFlow
 from jina.logging.profile import TimeContext
+from jina.types.document.generators import from_ndarray
 from jina.types.request import Response
 from tests import validate_callback
 
@@ -41,7 +42,7 @@ async def test_run_async_flow(restful, mocker):
     r_val = mocker.Mock()
     with AsyncFlow(restful=restful).add() as f:
         async for r in f.index(
-            DocumentArray.from_ndarray(np.random.random([num_docs, 4])), on_done=r_val
+            from_ndarray(np.random.random([num_docs, 4])), on_done=r_val
         ):
             assert isinstance(r, Response)
     validate_callback(r_val, validate)
@@ -90,7 +91,7 @@ async def run_async_flow_5s(restful):
 
     with AsyncFlow(restful=restful).add(uses=Wait5s) as f:
         async for r in f.index(
-            DocumentArray.from_ndarray(np.random.random([num_docs, 4])),
+            from_ndarray(np.random.random([num_docs, 4])),
             on_done=validate,
         ):
             assert isinstance(r, Response)
@@ -138,7 +139,7 @@ async def test_run_async_flow_other_task_concurrent(restful):
 @pytest.mark.parametrize('restful', [False])
 async def test_return_results_async_flow(return_results, restful):
     with AsyncFlow(restful=restful, return_results=return_results).add() as f:
-        async for r in f.index(DocumentArray.from_ndarray(np.random.random([10, 2]))):
+        async for r in f.index(from_ndarray(np.random.random([10, 2]))):
             assert isinstance(r, Response)
 
 
@@ -154,7 +155,6 @@ async def test_return_results_async_flow_crud(return_results, restful, flow_api)
 
 @pytest.mark.asyncio
 async def test_async_flow_empty_data():
-
     from jina import Executor, requests
 
     class MyExec(Executor):
