+++ tests/unit/flow/test_flow.py
@@ -3,12 +3,13 @@ import os
 import numpy as np
 import pytest
 
-from jina import Flow, Document, DocumentArray
+from jina import Flow, Document
 from jina.enums import SocketType, FlowBuildLevel
 from jina.excepts import RuntimeFailToStart
 from jina.executors import BaseExecutor
 from jina.helper import random_identity
 from jina.peapods.pods import BasePod
+from jina.types.document.generators import from_ndarray
 from jina.types.request import Response
 from tests import random_docs, validate_callback
 
@@ -457,7 +458,7 @@ def test_flow_needs_all(restful):
     assert f._pod_nodes['r2'].needs == {'p3', 'r1'}
 
     with f:
-        f.index(DocumentArray.from_ndarray(np.random.random([10, 10])))
+        f.index(from_ndarray(np.random.random([10, 10])))
 
     f = (
         Flow(restful=restful)
@@ -472,7 +473,7 @@ def test_flow_needs_all(restful):
     assert f._pod_nodes['p4'].needs == {'r2'}
 
     with f:
-        f.index(DocumentArray.from_ndarray(np.random.random([10, 10])))
+        f.index(from_ndarray(np.random.random([10, 10])))
 
 
 def test_flow_with_pod_envs():
@@ -508,7 +509,7 @@ def test_flow_with_pod_envs():
 @pytest.mark.parametrize('restful', [False, True])
 def test_return_results_sync_flow(return_results, restful):
     with Flow(restful=restful, return_results=return_results).add() as f:
-        r = f.index(DocumentArray.from_ndarray(np.random.random([10, 2])))
+        r = f.index(from_ndarray(np.random.random([10, 2])))
         if return_results:
             assert isinstance(r, list)
             assert isinstance(r[0], Response)
