+++ tests/unit/flow/test_flow_except.py
@@ -1,8 +1,9 @@
 import numpy as np
 import pytest
 
-from jina import Flow, Executor, requests, Document, DocumentArray
+from jina import Flow, Executor, requests, Document
 from jina.proto import jina_pb2
+from jina.types.document.generators import from_ndarray
 from tests import validate_callback
 
 
@@ -185,7 +186,7 @@ def test_flow_on_callback(restful):
 
     with f:
         f.index(
-            DocumentArray.from_ndarray(np.random.random([10, 10])),
+            from_ndarray(np.random.random([10, 10])),
             on_done=f1,
             on_error=f2,
             on_always=f3,
@@ -217,7 +218,7 @@ def test_flow_on_error_callback(restful):
 
     with f:
         f.index(
-            DocumentArray.from_ndarray(np.random.random([10, 10])),
+            from_ndarray(np.random.random([10, 10])),
             on_done=f1,
             on_error=f2,
             on_always=f3,
